2023/python/requirements.txt
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
PyYAML==6.0.1
